Saving relative+absolute positioning solution. Not sure if this is the best way to solve the problem of converting flowchart boxes I intend to be HTML inputs. I looked for something like svg -> html converter to no avail.
This feels very much like a 2004 solution. The problem with this approach is that to change what the flowchart looks like you then need to change the html input box location. Workaround for this is to 'just make sure the flowchart svg is done before moving to html positioning' but I feel like a programmatic workflow exists out there somewhere.

Perhaps a solution is to find an editor that will produce html with arrows from an easy to use gui like I used draw.io for building the svg image.