First draft of importKey() tooltip.
I'm imagining 2 different types of tooltips - 1 for SubtleCrypto functions and 1 for our proprietary functions.

For the SubtleCrypo functions I think I like the exhaustive list of documentation links with an initial super concise description like a one-liner.

I'm not sure I like the idea of a block of text at the top of each function pane giving a description because I find myyself immediately getting sleepy/checked out when I go to the Apache (really sleepy) or MDN (better but still) documentation and am met with a similar block of text.

I like the idea of a dev site you can jump in and immediately be entering crypto things and running functions without the gatekeeper/distractor block of description text.

Perhaps a small (+) subscript button on the function title?